validaciones

+++ TP1/math.h
@@ -1,6 +1,6 @@
-void ingresarNumero(float*);
-float sumar(float,float);
-float restar(float,float);
-float multiplicar(float,float);
-float dividir(float,float);
-float calcularFactorial(float);
+void ingresarNumero(double*);
+double sumar(double,double);
+double restar(double,double);
+double multiplicar(double,double);
+double dividir(double,double);
+double calcularFactorial(double);
